Fix landing on a nasty again
It seems the last fix introduced the bug from the original version which can have Willy land into a solid wall.
So the true fix seems to be that if one of the tiles below Willy is harmful, if he is jumping (descending in this case) and the other tile is something he can move into, e.g. empty space or an item, then continue with the jump, else kill Willy.